wip: initial refactoring of tests
