+++ package.json
@@ -81,6 +81,7 @@
     "@express-rate-limit/tsconfig": "1.0.2",
     "@jest/globals": "29.7.0",
     "@types/express": "4.17.18",
+    "@types/jest": "29.5.5",
     "@types/supertest": "2.0.12",
     "body-parser": "1.20.2",
     "del-cli": "5.1.0",
@@ -101,11 +102,11 @@
     "prettier": true,
     "rules": {
       "@typescript-eslint/prefer-nullish-coalescing": [
-      	"error",
-      	{
-      		"ignoreConditionalTests": true
-      	}
-     	],
+        "error",
+        {
+          "ignoreConditionalTests": true
+        }
+      ],
       "@typescript-eslint/no-confusing-void-expression": 0,
       "@typescript-eslint/consistent-indexed-object-style": [
         "error",
