+++ source/index.ts
@@ -1,0 +1,8 @@
+// /source/index.ts
+// Export away!
+
+// Export all the types as named exports.
+export * from './types.js'
+
+// Export the slowDown function as well.
+export { default, default as slowDown } from './express-slow-down.js'
