+++ test/middleware-test.ts
@@ -1,17 +1,13 @@
-import EventEmitter from 'node:events'
-import {
-	describe,
-	expect,
-	beforeEach,
-	afterEach,
-	jest,
-	it,
-} from '@jest/globals'
-import slowDown from '../source/express-slow-down'
-import { expectDelay, expectNoDelay } from './helpers/requests'
-import { MockStore } from './helpers/mock-store'
+// /test/middleware-test.ts
+// Tests the middleware by passing in various options to it
 
-describe('request skipping', () => {
+import EventEmitter from 'node:events'
+import { jest } from '@jest/globals'
+import slowDown from '../source/index.js'
+import { expectDelay, expectNoDelay } from './helpers/requests.js'
+import { MockStore } from './helpers/mock-stores.js'
+
+describe('middleware behaviour', () => {
 	beforeEach(() => {
 		jest.useFakeTimers()
 		jest.spyOn(global, 'setTimeout')
@@ -21,7 +17,47 @@ describe('request skipping', () => {
 		jest.restoreAllMocks()
 	})
 
-	// Skip
+	it('should allow delayAfter to be a function', async () => {
+		const instance = slowDown({
+			delayAfter: () => 2,
+			delayMs: 99,
+		})
+		await expectNoDelay(instance)
+		await expectNoDelay(instance)
+		await expectDelay(instance, 99)
+	})
+
+	it('should allow delayMs to be a function', async () => {
+		const instance = slowDown({
+			delayAfter: 1,
+			delayMs: () => 99,
+		})
+		await expectNoDelay(instance)
+		await expectDelay(instance, 99)
+	})
+
+	it('should allow maxDelayMs to be a function', async () => {
+		const instance = slowDown({
+			delayAfter: 1,
+			delayMs: (used: number) => (used - 1) * 100,
+			maxDelayMs: () => 200,
+		})
+		await expectNoDelay(instance)
+		await expectDelay(instance, 100)
+		await expectDelay(instance, 200)
+		await expectDelay(instance, 200)
+	})
+
+	it('should allow a custom key generator', async () => {
+		const keyGenerator = jest.fn() as any
+		const instance = slowDown({
+			delayAfter: 1,
+			keyGenerator,
+		})
+
+		await expectNoDelay(instance)
+		expect(keyGenerator).toHaveBeenCalled()
+	})
 
 	it('should allow a custom skip function', async () => {
 		const skip = jest
@@ -39,8 +75,6 @@ describe('request skipping', () => {
 		await expectNoDelay(instance)
 		expect(skip).toHaveBeenCalledTimes(2)
 	})
-
-	// SkipSuccessfulRequests
 
 	it('should decrement hits with success response and skipSuccessfulRequests', async () => {
 		const request = {}
@@ -74,8 +108,6 @@ describe('request skipping', () => {
 		res.emit('finish')
 		expect(store.decrementWasCalled).toBeFalsy()
 	})
-
-	// SkipFailedRequests
 
 	it('should not decrement hits with success response and skipFailedRequests', async () => {
 		const request = {}
@@ -139,5 +171,21 @@ describe('request skipping', () => {
 
 		res.emit('error')
 		expect(store.decrementWasCalled).toBeTruthy()
+	})
+
+	it('should augment the req object with info about the slowdown status', async () => {
+		const request: any = {}
+		const instance = slowDown({
+			delayAfter: 2,
+			windowMs: 1000,
+		})
+		await expectNoDelay(instance, request)
+		expect(request.slowDown).toMatchObject({
+			current: 1,
+			delay: 0,
+			limit: 2,
+			remaining: 1,
+			resetTime: expect.any(Date),
+		})
 	})
 })
