--- test/options.test.ts
@@ -1,25 +1,0 @@
-import {
-	describe,
-	expect,
-	beforeEach,
-	afterEach,
-	jest,
-	it,
-} from '@jest/globals'
-import slowDown from '../source/express-slow-down'
-
-describe('options', () => {
-	beforeEach(() => {
-		jest.useFakeTimers()
-	})
-	afterEach(() => {
-		jest.useRealTimers()
-		jest.restoreAllMocks()
-	})
-
-	it('should not modify the options object passed', () => {
-		const options = {}
-		slowDown(options)
-		expect(options).toStrictEqual({})
-	})
-})
